Task 2(Multi-Modal-Chatbot)/requirements.txt
streamlit
Pillow
requests
google-generativeai
scikit-learn
pandas
seaborn
matplotlib
joblib
imbalanced-learn